pyproject.toml
[tool.poetry]
name = "send"
version = "0.1.0"
description = ""
authors = ["Wongsathorn <mic.mond@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "*"
django-environ = "^0.4.5"
django-background-tasks = "^1.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
